+++ src/main/java/bnorbert/auction/domain/Bid.java
@@ -9,7 +9,7 @@ public class Bid {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private double amount;
+    private long amount;
     private Instant createdDate;
 
     @ManyToOne(fetch = FetchType.LAZY)
@@ -29,11 +29,11 @@ public class Bid {
         this.id = id;
     }
 
-    public double getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
