+++ src/main/java/bnorbert/auction/domain/Transaction.java
@@ -8,7 +8,7 @@ public class Transaction {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private double amount;
+    private long amount;
     private String description;
     private TransactionType transactionType;
     private String transferToAccountId;
@@ -35,11 +35,11 @@ public class Transaction {
         this.account = account;
     }
 
-    public double getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
