+++ src/main/java/bnorbert/auction/transfer/home/HomeResponse.java
@@ -14,7 +14,7 @@ public class HomeResponse {
     private String garageYearBuilt;
     private int garageCars;
     private int garageArea;
-    private double startingPrice;
+    private long startingPrice;
     private Long roomId;
     private Long timeSlotId;
     private DayOfWeek dayOfWeek;
@@ -101,11 +101,11 @@ public class HomeResponse {
         this.garageArea = garageArea;
     }
 
-    public double getStartingPrice() {
+    public long getStartingPrice() {
         return startingPrice;
     }
 
-    public void setStartingPrice(double startingPrice) {
+    public void setStartingPrice(long startingPrice) {
         this.startingPrice = startingPrice;
     }
